% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entertain_me.R
\name{entertain_me}
\alias{entertain_me}
\title{entertain_me}
\usage{
entertain_me()
}
\value{
A meme on the \code{Viewer} Pane
}
\description{
Displays memes on the \code{Viewer} pane
}
\examples{
entertain_me()
}
